Week2/Day5/Blog/requirements.txt
Flask
psycopg[binary]
translate
gTTS
python-dotenv